@@ -2,6 +2,11 @@
 
 namespace BenRowan\VCsvStream\Tests\Small\Generators;
 
+use Faker\Factory;
+use PHPUnit\Framework\Attributes\DataProvider;
+use PHPUnit\Framework\Attributes\Test;
+use Iterator;
+use OverflowException;
 use BenRowan\VCsvStream\Generators\FakerValue;
 use BenRowan\VCsvStream\Tests\Assets\AbstractTestCase;
 use Faker;
@@ -12,17 +17,13 @@ class FakerValueTest extends AbstractTestCase
 
     public function setUp(): void
     {
-        $this->faker = Faker\Factory::create();
+        $this->faker = Factory::create();
         $this->faker->seed(1);
     }
 
-    /**
-     * @test
-     *
-     * @param string $property
-     *
-     * @dataProvider iCanGenerateValuesDataProvider
-     */
+    
+    #[DataProvider('iCanGenerateValuesDataProvider')]
+    #[Test]
     public function iCanGenerateValues(string $property): void
     {
         $fakerValue = new FakerValue($this->faker, $property);
@@ -30,23 +31,19 @@ class FakerValueTest extends AbstractTestCase
         $this->assertNotEmpty($fakerValue->generate());
     }
 
-    public function iCanGenerateValuesDataProvider(): array
+    public static function iCanGenerateValuesDataProvider(): Iterator
     {
-        return [
-            ['text'],
-            ['randomNumber'],
-            ['randomFloat'],
-            ['email']
-        ];
+        yield ['text'];
+        yield ['randomNumber'];
+        yield ['randomFloat'];
+        yield ['email'];
     }
 
-    /**
-     * @test
-     */
+    #[Test]
     public function iCanGenerateUniqueValues(): void
     {
         // Faker throws this when it runs out of unique values.
-        $this->expectException(\OverflowException::class);
+        $this->expectException(OverflowException::class);
 
         $fakerValue = new FakerValue($this->faker, 'randomDigitNotNull', true);
 
