@@ -4,22 +4,17 @@ namespace BenRowan\VCsvStream\Generators;
 
 final class FixedValue implements GeneratorInterface
 {
-    private $value;
-
     /**
      * Create a fixed value generator.
      *
      * @param mixed $value Anything which can be cast to a string.
      */
-    public function __construct($value)
+    public function __construct(private readonly mixed $value)
     {
-        $this->value = $value;
     }
 
     /**
      * Generate a fixed value.
-     *
-     * @return string
      */
     public function generate(): string
     {
