@@ -2,6 +2,8 @@
 
 namespace BenRowan\VCsvStream\Renderers;
 
+use BenRowan\VCsvStream\Stream\ConfigInterface;
+use BenRowan\VCsvStream\Stream\StateInterface;
 use BenRowan\VCsvStream\Exceptions\VCsvStreamException;
 use BenRowan\VCsvStream\Stream;
 use BenRowan\VCsvStream\Rows\Header as HeaderRow;
@@ -11,14 +13,11 @@ class Header extends AbstractRowRenderer
     /**
      * Renders the current header data as a CSV row string.
      *
-     * @param Stream\ConfigInterface $config
-     * @param Stream\StateInterface $streamState
      *
-     * @return string
      *
      * @throws VCsvStreamException
      */
-    public function render(Stream\ConfigInterface $config, Stream\StateInterface $streamState): string
+    public function render(ConfigInterface $config, StateInterface $streamState): string
     {
         if (! $streamState->hasHeader()) {
             throw new VCsvStreamException(
