@@ -2,9 +2,11 @@
 
 namespace BenRowan\VCsvStream\Renderers;
 
+use BenRowan\VCsvStream\Stream\ConfigInterface;
+use BenRowan\VCsvStream\Stream\StateInterface;
 use BenRowan\VCsvStream\Stream;
 
 interface RowRendererInterface
 {
-    public function render(Stream\ConfigInterface $config, Stream\StateInterface $streamState): string;
+    public function render(ConfigInterface $config, StateInterface $streamState): string;
 }
